video_detect/requirements.txt
torch
transformers
pillow
gradio
opencv-python
django
django-filter
djangorestframework
markdown
yt_dlp